Remove automatic updating line from OpenSearch description
This is causing problems installing the search provider now. If
it uses the domain where the description was installed from,
then it will 404 because that url doens't exist. If it's relative
to the search domain, it will update the search provider with the
non-hidden service url. Neither option is desirable so we'll just
remove it.